Move Assignment files to assignment folder
